+++ views/auth/register.view.php
@@ -4,29 +4,48 @@
     <div class="sm:mx-auto sm:w-full sm:max-w-sm col-span-full">
         <h2 class="my-10 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900">Add a new author</h2>
     </div>
-    <form class="min-w-[40%]">
+    <form method="POST" action="" class="min-w-[40%]" novalidate>
         <div class="space-y-12">
             <div class="border-b border-gray-900/10 pb-12">
                 <div class="grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
                     <div class="sm:col-span-3">
-                        <label for="first-name" class="block text-sm font-medium leading-6 text-gray-900">First name</label>
+                        <label for="name" class="block text-sm font-medium leading-6 text-gray-900">First name</label>
                         <div class="mt-2">
-                            <input type="text" name="first-name" id="first-name" autocomplete="given-name" class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                            <input type="text" name="name" id="name" value="<?= $_POST['name'] ?? '' ?>" autocomplete="name" class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
                         </div>
+                        <?php if (isset($errors['name'])) : ?>
+                            <span class="text-red-500 text-xs font-medium"><?= $errors['name'] ?></span>
+                        <?php endif ?>
                     </div>
 
                     <div class="sm:col-span-3">
                         <label for="username" class="block text-sm font-medium leading-6 text-gray-900">Username</label>
                         <div class="mt-2">
-                            <input id="username" name="username" type="text" autocomplete="username" class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                            <input id="username" name="username" type="text" value="<?= $_POST['username'] ?? '' ?>" autocomplete="username" class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
                         </div>
+                        <?php if (isset($errors['username'])) : ?>
+                            <span class="text-red-500 text-xs font-medium"><?= $errors['username'] ?></span>
+                        <?php endif ?>
                     </div>
 
-                    <div class="sm:col-span-4">
+                    <div class="sm:col-span-3">
                         <label for="email" class="block text-sm font-medium leading-6 text-gray-900">Email address</label>
                         <div class="mt-2">
-                            <input id="email" name="email" type="email" autocomplete="email" class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                            <input id="email" name="email" type="email" value="<?= $_POST['email'] ?? '' ?>" autocomplete="email" class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
                         </div>
+                        <?php if (isset($errors['email'])) : ?>
+                            <span class="text-red-500 text-xs font-medium"><?= $errors['email'] ?></span>
+                        <?php endif ?>
+                    </div>
+
+                    <div class="sm:col-span-3">
+                        <label for="password" class="block text-sm font-medium leading-6 text-gray-900">Password</label>
+                        <div class="mt-2">
+                            <input id="password" name="password" value="<?= $_POST['password'] ?? '' ?>" type="password" autocomplete="password" class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6">
+                        </div>
+                        <?php if (isset($errors['password'])) : ?>
+                            <span class="text-red-500 text-xs font-medium"><?= $errors['password'] ?></span>
+                        <?php endif ?>
                     </div>
 
                 </div>